Now TargetYaw can be edited anywhere
- In OpenDoor.h right before TargetYaw, called UPROPERTY(EditAnywhere),
which allows whatever comes after it, in our case TargetYaw to be edited
from anywhere. In one place, OpenDoor.cpp, we edit TargetYaw in general
for both to be set to 90.f degrees. Then in Unreal Engine Scene, we
edited the 2nd door's TargetYaw to be 30.f degrees, so when game is played,
the first door opens at 90.f degrees andd the 2nd door opens at 30.f deg